# openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Notifuse API",
    "description": "API for Notifuse - a transactional email and newsletter management platform",
    "version": "1.0.0",
    "contact": {
      "name": "Notifuse Support",
      "url": "https://www.notifuse.com/support",
      "email": "hello@notifuse.com"
    }
  },
  "servers": [
    {
      "url": "https://{notifuseDomain}",
      "description": "Customer-specific Notifuse API server",
      "variables": {
        "notifuseDomain": {
          "description": "Your unique Notifuse domain",
          "default": "demo.notifuse.com"
        }
      }
    }
  ],
  "paths": {
...
    "/api/broadcasts.list": {
      "get": {
        "summary": "List broadcasts",
        "description": "Retrieves a list of broadcasts with pagination and optional filtering by status. Supports fetching template details for each variation.",
        "operationId": "listBroadcasts",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "workspace_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "The ID of the workspace",
            "example": "ws_1234567890"
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "draft",
                "scheduled",
                "sending",
                "paused",
                "sent",
                "cancelled",
                "failed",
                "testing",
                "test_completed",
                "winner_selected"
              ]
            },
            "description": "Filter broadcasts by status"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 50
            },
            "description": "Maximum number of broadcasts to return",
            "example": 50
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 0
            },
            "description": "Number of broadcasts to skip for pagination",
            "example": 0
          },
          {
            "name": "with_templates",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false
            },
            "description": "Include full template details for each variation",
            "example": false
          }
        ],
        "responses": {
          "200": {
            "description": "List of broadcasts retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BroadcastListResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad request - validation failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "error": "workspace_id is required"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - invalid or missing authentication token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "error": "Failed to list broadcasts"
                }
              }
            }
          }
        }
      }
    },
    "/api/broadcasts.get": {
      "get": {
        "summary": "Get a broadcast",
        "description": "Retrieves a single broadcast by ID. Optionally includes template details for each variation.",
        "operationId": "getBroadcast",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "workspace_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "The ID of the workspace",
            "example": "ws_1234567890"
          },
          {
            "name": "id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "The ID of the broadcast",
            "example": "broadcast_12345"
          },
          {
            "name": "with_templates",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false
            },
            "description": "Include full template details for each variation",
            "example": false
          }
        ],
        "responses": {
          "200": {
            "description": "Broadcast retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "broadcast": {
                      "$ref": "#/components/schemas/Broadcast"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad request - validation failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - invalid or missing authentication token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Broadcast not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "error": "Broadcast not found"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/broadcasts.create": {
      "post": {
        "summary": "Create a broadcast",
        "description": "Creates a new broadcast campaign in draft status.\n\n**Important:** Creating a broadcast does not schedule or send it. After creating the broadcast,\nyou must call the `/api/broadcasts.schedule` endpoint to schedule or send it immediately.\n",
        "operationId": "createBroadcast",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateBroadcastRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Broadcast created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "broadcast": {
                      "$ref": "#/components/schemas/Broadcast"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad request - validation failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "examples": {
                  "missingName": {
                    "value": {
                      "error": "name is required"
                    }
                  },
                  "missingList": {
                    "value": {
                      "error": "list is required"
                    }
                  },
                  "invalidTestSettings": {
                    "value": {
                      "error": "at least 2 variations are required for A/B testing"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - invalid or missing authentication token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "error": "Failed to create broadcast"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "API token for authentication"
      }
    },
    "schemas": {
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "description": "Error message"
          }
        },
        "required": [
          "error"
        ]
      },
...
      "AudienceSettings": {
        "type": "object",
        "required": [
          "list"
        ],
        "properties": {
          "list": {
            "type": "string",
            "description": "List ID to send to",
            "example": "newsletter"
          },
          "segments": {
            "type": "array",
            "description": "Optional segment IDs to filter recipients",
            "items": {
              "type": "string"
            },
            "example": [
              "premium_users"
            ]
          },
          "exclude_unsubscribed": {
            "type": "boolean",
            "description": "Whether to exclude unsubscribed contacts",
            "example": true
          }
        }
      },
...
      "CreateBroadcastRequest": {
        "type": "object",
        "required": [
          "workspace_id",
          "name",
          "audience"
        ],
        "properties": {
          "workspace_id": {
            "type": "string",
            "description": "The ID of the workspace",
            "example": "ws_1234567890"
          },
          "name": {
            "type": "string",
            "description": "Name of the broadcast",
            "example": "Spring Newsletter 2024",
            "maxLength": 255
          },
          "audience": {
            "$ref": "#/components/schemas/AudienceSettings"
          },
          "test_settings": {
            "$ref": "#/components/schemas/BroadcastTestSettings"
          },
          "tracking_enabled": {
            "type": "boolean",
            "description": "Enable click and open tracking",
            "example": true
          },
          "utm_parameters": {
            "$ref": "#/components/schemas/UTMParameters"
          },
          "data_feed": {
            "$ref": "#/components/schemas/DataFeedSettings"
          },
          "metadata": {
            "type": "object",
            "additionalProperties": true,
            "description": "Custom metadata for the broadcast"
          }
        }
      },
...
      "BroadcastListResponse": {
        "type": "object",
        "properties": {
          "broadcasts": {
            "type": "array",
            "description": "List of broadcasts",
            "items": {
              "$ref": "#/components/schemas/Broadcast"
            }
          },
          "total_count": {
            "type": "integer",
            "description": "Total number of broadcasts matching the query",
            "example": 45
          }
        }
      },
...
    }
  }
}